Ran - CR and shit
