+++ application/src/strategy/ComparisonTab.tsx
@@ -4,6 +4,7 @@ import BoxChart from "./charts/BoxChart";
 import { Match } from "../utils/Match";
 import { fetchAllTeamMatches, fetchPaticularTeamMatches } from "../utils/Fetches";
 import { useRecent } from "../components/TeamPicker";
+import { FRCTeamArray } from "../utils/Utils";
 
 interface FieldOption {
   name: string;
@@ -44,64 +45,60 @@ const fieldOptions: FieldOption[] = [
     getData: (match) =>
       match.teleReefPick.algea.netScore + match.autoReefPick.algea.netScore,
   },
-
   {
     name: "Processor",
     getData: (match) =>
       match.teleReefPick.algea.processor + match.autoReefPick.algea.processor,
   },
-
   {
     name: "Team Number",
-    getData: (match) => 
-      match.teamNumber
+    getData: (match) => match.teamNumber,
   },
   {
     name: "Qual",
-    getData: (match) =>
-       match.qual 
+    getData: (match) => match.qual,
   },
 ];
 
-const teamNumbers = [
-  {id: "1", value: 1937},
-  {id: "2", value: 4590},
-]
+const teamNumbers = FRCTeamArray
 
 const ComparisonTab: React.FC = () => {
   const [teams, setTeams] = useState<TeamData[]>([]);
   const [recency, setRecency] = useState<number>(0);
-  const[checkedList, setCheckedList] = useState<number[]>([]);
+  const [checkedList, setCheckedList] = useState<number[]>([]);
+  console.log(teamNumbers)
 
-  const handleSelect = (event) =>{
-    const value = event.target.value
-    const isChecked = event.target.checked
-      if(isChecked){
-        const addedList = [...checkedList, value]
-        setCheckedList(addedList)
-      }
-      else{
-        const filteredList = checkedList.filter((item)=> item!==value)
-        setCheckedList(filteredList)
-      }
-  }
+  const handleSelect = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const value = Number(event.target.value);
+    const isChecked = event.target.checked;
+
+    setCheckedList((prevList) =>
+      isChecked ? [...prevList, value] : prevList.filter((item) => item !== value)
+    );
+  };
 
   useEffect(() => {
     async function updateParticularTeams() {
-      console.log(
-        setTeams(Object.values(await fetchPaticularTeamMatches(checkedList)).map(
-          (teamMatches) => new TeamData(useRecent(teamMatches, recency))
-        )));
+      if (checkedList.length === 0) {
+        setTeams([]);
+        return;
+      }
+
+      const fetchedTeams = await fetchPaticularTeamMatches(checkedList);
+      const teamObjects = Object.values(fetchedTeams).map(
+        (teamMatches) => new TeamData(useRecent(teamMatches, recency))
+      );
+
+      setTeams(teamObjects);
     }
+
     updateParticularTeams();
-  });
+  }, [checkedList, recency]);
 
   const [field, setField] = useState<string>(fieldOptions[0].name);
 
   const getBoxData = (team: TeamData | undefined) => {
-    if (!team) {
-      return {};
-    }
+    if (!team) return {};
     const chosenOption =
       fieldOptions.find((option) => option.name === field) || fieldOptions[0];
     return {
@@ -109,23 +106,19 @@ const ComparisonTab: React.FC = () => {
     };
   };
 
-  const teamsData = useMemo(() => {
-    return teams.map((team) => getBoxData(team)) || [];
-  }, [teams, field]);
+  const teamsData = useMemo(() => teams.map((team) => getBoxData(team)) || [], [teams, field]);
 
   return (
     <>
       <div className="flex flex-col items-center">
         <br />
 
-        <select
-          className="my-5"
-          onChange={(event) => setField(event.currentTarget.value)}
-        >
+        <select className="my-5" onChange={(event) => setField(event.currentTarget.value)}>
           {fieldOptions.map((option) => (
-            <option>{option.name}</option>
+            <option key={option.name}>{option.name}</option>
           ))}
         </select>
+
         <label htmlFor="recency">Recency</label>
         <input
           type="number"
@@ -134,7 +127,8 @@ const ComparisonTab: React.FC = () => {
           onChange={(event) => setRecency(parseInt(event.target.value))}
           min={1}
         />
-        <div className="w-96 ">
+
+        <div className="w-96">
           <BoxChart
             data={Object.assign({}, ...teamsData)}
             xName={"Team"}
@@ -144,19 +138,24 @@ const ComparisonTab: React.FC = () => {
           />
         </div>
       </div>
-      <div className="teams">
-        {teamNumbers.map((item)=>{
-          return(
-            <input 
-            type="checkbox"
-            name="team number"
-            id={item.id}
-            value={item.value}
-            onChange={handleSelect} 
+      <div className="teams" style={{display: "grid", gap: "0.55em", gridTemplateColumns: "repeat(5, 1fr)", gridTemplateRows: "repeat(5, 1fr)",}}>
+        {teamNumbers.map((item) => (
+          <label className="containerComp" key={item.id} style={{ display: "flex", alignItems: "center", gap: "0.3em" }}>
+            <input
+              type="checkbox"
+              name="team number"
+              id={item.id}
+              value={item.value}
+              onChange={handleSelect}
+              style={{
+                width: "1.2em",
+                height: "1.2em",
+              }}
             />
-          )
-        })}
-      </div>
+            <span className="checkmarkComp"></span>
+          {item.id}</label>
+        ))}
+        </div>
     </>
   );
 };
